+++ codex/urls/converters.py
@@ -1,0 +1,8 @@
+"""Custom url converters."""
+from django.urls.converters import StringConverter
+
+
+class GroupConverter(StringConverter):
+    """Only accept valid browser groups."""
+
+    regex = "[rpisvcf]"