+++ frontend/src/components/auth/auth-menu.vue
@@ -5,7 +5,8 @@
       <v-list-item ripple @click="logout">
         <v-list-item-content>
           <v-list-item-title>
-            <h3>Logout {{ user.username }}</h3>
+            <v-icon>{{ mdiLogout }}</v-icon
+            >Logout {{ user.username }}
           </v-list-item-title>
         </v-list-item-content>
       </v-list-item>
@@ -16,6 +17,7 @@
 </template>
 
 <script>
+import { mdiLogout } from "@mdi/js";
 import { mapActions, mapState } from "pinia";
 
 import AuthChangePasswordDialog from "@/components/auth/auth-change-password-dialog.vue";
@@ -30,7 +32,7 @@ export default {
   },
   data: function () {
     return {
-      item: undefined,
+      mdiLogout,
     };
   },
   computed: {