+++ lint-backend.sh
@@ -32,7 +32,6 @@ if [ "$(uname)" = "Darwin" ]; then
     # Hadolint & shfmt are difficult to install on linux
     # shellcheck disable=2035
     hadolint *.Dockerfile
-    #shfmt -d -i 4 ./*.sh ./**/*.sh
     shellharden ./*.sh ./**/*.sh
     # subdirs aren't copied into docker builder
     # .env files aren't copied into docker